Rename SpawnMenu to TroopSpawnMenu to better differentiate from new PlayerSpawnMenu.
