+++ Alliance.Client/_Module/GUI/Prefabs/TroopSpawnMenu/Footer.xml
