+++ Alliance.Client/_Module/GUI/Prefabs/TroopSpawnMenu/FormationList.xml
