+++ Alliance.Client/_Module/GUI/Prefabs/TroopSpawnMenu/RecruitmentOptions.xml
