+++ Alliance.Client/_Module/GUI/Prefabs/TroopSpawnMenu/TroopGroup.xml
