+++ Alliance.Client/_Module/GUI/Prefabs/TroopSpawnMenu/TroopInfos.xml
