+++ Alliance.Client/_Module/GUI/Prefabs/TroopSpawnMenu/TroopList.xml
