+++ Alliance.Client/_Module/GUI/Prefabs/TroopSpawnMenu/TroopPreview.xml
