+++ Alliance.Client/_Module/GUI/Prefabs/TroopSpawnMenu/TroopSpawnMenu.xml
